refac: refact loginbutton
